refactor: Clean up unused imports and fix import paths
- Update credential service import paths from `../../../services` to `../../services` in OAuth credential files
- Remove unused `User` icon import from SettingsLayout component
- Remove unused `useMemo` hook import from NodeMarketplace component
- Remove unused `React` import from ExampleWorkflowCardWithTeam and TeamBadge components
- Remove unused `Play` icon import from ExampleWorkflowCardWithTeam component
- Fix parameter naming in ScheduledExecutionsList deleteJob function (prefix unused `workflowId` with underscore)
- Consolidate and correct node execution state references in AIAgentNode component
- Simplify NodeHeader prop structure by consolidating icon configuration
- Standardize NodeContextMenu prop naming for consistency
- Improve code maintainability by removing dead code and correcting import paths
