+++ frontend/src/components/workflow/nodes/NodeLabel.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+
 
 interface NodeLabelProps {
   label: string
